change the salary type to DECIMAL

+++ models/staff.js
@@ -24,7 +24,7 @@ module.exports = (sequelize, DataTypes) => {
     country_code: DataTypes.INTEGER,
     phone: DataTypes.STRING,
     salary_type_txid: DataTypes.INTEGER,
-    salary: DataTypes.BIGINT,
+    salary: DataTypes.DECIMAL,
     cycle_start_day: DataTypes.INTEGER,
     cycle_start_date: DataTypes.INTEGER,
     daily_shift_duration: DataTypes.TIME,
